チェック対応 	modified:   app/controllers/artists_controller.rb 	modified:   app/controllers/home_controller.rb 	modified:   app/controllers/twitters_controller.rb 	modified:   app/helpers/artists_helper.rb 	new file:   app/javascript/application.js 	new file:   app/javascript/controllers/application.js 	new file:   app/javascript/custom/hoge.js 	modified:   app/models/artist.rb 	modified:   app/models/concerns/pxv.rb 	modified:   app/models/concerns/twt.rb 	modified:   app/models/concerns/url_txt_reader.rb 	modified:   app/views/artists/_form.html.erb 	modified:   app/views/artists/_pics.html.erb 	modified:   app/views/artists/_stat.html.erb 	modified:   app/views/artists/index.html.erb 	modified:   app/views/artists/twt_index.html.erb 	modified:   app/views/home/index.html.erb 	modified:   app/views/twitters/_form.html.erb 	modified:   app/views/twitters/_twt_pic_list.erb 	modified:   app/views/twitters/index.html.erb 	new file:   config/application.rb 	new file:   config/environments/production.rb 	new file:   config/importmap.rb

+++ app/models/concerns/pxv.rb
@@ -4,6 +4,7 @@ module Pxv
     extend ActiveSupport::Concern
     
     PXV_DIRLIST_PATH = "public/pxv/dirlist.txt"
+    PXV_ARCHIVE_DIR_PATH = "public/f_dl/PxDl/"
 
     def self.pxv_user_url(pxvid)
         %!https://www.pixiv.net/users/#{pxvid}!
@@ -16,6 +17,10 @@ module Pxv
     def self.get_path_from_dirlist(pxvid)
         rpath = []
         txtpath = Rails.root.join(PXV_DIRLIST_PATH).to_s
+        unless File.exist? txtpath
+            return rpath
+        end
+
         File.open(txtpath) { |file|
             while line  = file.gets
                 #if line.include?(search_str)
@@ -30,7 +35,7 @@ module Pxv
     def self.get_pathlist(pxvid)
         path_list = []
 
-        current_work_dir_root = Rails.root.join("public/f_dl/PxDl/").to_s + "*/"
+        current_work_dir_root = Rails.root.join(PXV_ARCHIVE_DIR_PATH).to_s + "*/"
         Dir.glob(current_work_dir_root).each do |path|
             if File.basename(path) =~ /\(#{pxvid}\)/
                 puts %!path=#{path}!
